requirements.txt
python-dotenv
strands
pytesseract 
pillow
